Moved Location into it's own file
