+++ location/src/androidMain/kotlin/com/example/location/data/repository/AndroidLocationProvider.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import android.location.LocationManager
 import androidx.annotation.RequiresPermission
 import com.example.location.domain.model.LocationResult
-import com.example.shared.data.event.Location
+import com.example.shared.data.model.Location
 
 class AndroidLocationProvider(
     private val context: Context
