+++ location/src/androidMain/kotlin/com/example/location/platform/AndroidBackgroundLocationService.kt
@@ -7,7 +7,7 @@ import android.location.LocationListener
 import android.location.LocationManager
 import android.os.IBinder
 import com.example.shared.data.event.InPlayEvent
-import com.example.shared.data.event.Location
+import com.example.shared.data.model.Location
 import com.example.shared.platform.Logger
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
