+++ shared/src/commonMain/kotlin/com/example/shared/data/entity/LocationEntity.kt
@@ -2,7 +2,7 @@ package com.example.shared.data.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.example.shared.data.event.Location
+import com.example.shared.data.model.Location
 
 @Entity(tableName = "locations")
 data class LocationEntity(
